pyproject.toml
[project]
name = "invoice-extracter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.13",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
